chatbot/requirements.txt
# ================================
# Core Framework
# ================================
Django>=4.2,<5.0
djangorestframework>=3.14

# ================================
# Environment & Utilities
# ================================
python-dotenv>=1.0
requests>=2.31

# ================================
# Frontend & UI
# ================================
django-tailwind>=3.6
django-browser-reload>=1.12
django-crispy-forms>=2.1
crispy-tailwind>=0.5.0
widget_tweaks>=1.4.12
Pillow>=10.0

# ================================
# AI & Machine Learning
# ================================
google-genai>=0.2.0

# ================================
# NLP & Machine Learning
# ================================
nltk>=3.8
spacy>=3.7
scikit-learn>=1.3
numpy>=1.26

# ================================
# Database (optional for production)
# ================================
psycopg2-binary>=2.9
